requirements.txt
# ==============================================================================
# INSTALLATION INSTRUCTIONS
# ==============================================================================
# 1. Install PyTorch FIRST (Choose one based on your machine):
#
#    FOR LOCAL CPU:
#    pip install torch==2.11.0 torchvision==0.26.0 --index-url https://download.pytorch.org/whl/cpu
#
#    FOR LAB SERVER (A100 GPU):
#    pip install torch==2.11.0 torchvision==0.26.0 --index-url https://download.pytorch.org/whl/cu118
#    (Check server CUDA version with 'nvidia-smi' and use cu121 if needed)
#
# 2. Install the rest of the dependencies:
#    pip install -r requirements.txt
# ==============================================================================


# Medical Imaging & MONAI
monai==1.5.2
nibabel==5.4.2

# Visualization
matplotlib==3.10.8
tensorboard==2.20.0

# Environments
python-dotenv==1.2.2

# For logging stats
psutil==6.0.0
tqdm==4.66.5
